requirements.txt
flask>=3.0,<4
selenium>=4.14,<5
pillow>=10.0,<11
gunicorn>=21.2,<22
